- При выгрузке в ТЗ имена колонок задаются как есть, без замены символов. - Устранены ворнинги в MSVC x64. - Обновлён номер версии. - Актуализированы тесты в 1С.

+++ src/v8sqlite_addin.h
@@ -22,7 +22,7 @@ public:
         lstringu<40> v{ssa{P_VERSION}};
         value.vt = VTYPE_PWSTR;
         value.pwstrVal = copyText(v);
-        value.wstrLen  = v.length();
+        value.wstrLen  = (int)v.length();
         return true;
     }
